Create PolicyManager during bootstrap, allowing us to share initialization
